Add option to add a preset for the current date in rulesets
